проект MATRIX

+++ Vector/src/ru/academit/novikov/vector/Vector.java
@@ -93,7 +93,7 @@ public class Vector {
         return this;
     }
 
-    public void multiplicationByScalar(int scalar) {
+    public void multiplicationByScalar(double scalar) {
         for (int i = 0; i < this.components.length; i++) {
             this.components[i] = scalar * this.components[i];
         }
